requirements.txt
fastapi
uvicorn
supabase
python-dotenv
pandas
requests
matplotlib
streamlit
